requirements.txt
unstructured[pdf] 
python-dotenv
google-generativeai
opensearch-py
langchain
gradio
pymupdf
requests